Revert "Merge remote-tracking branch 'origin/mpi-threadpool-debugging' into wedge_metis"
This reverts commit 430513462f51e8c179382f5cfb3ea7a7cac15dab, reversing
changes made to db9542823f38c7fac8c225383b5823c5c107bf7a.

+++ src/distributed_io.c
@@ -169,8 +169,7 @@ void write_distributed_array(
     h_err = H5Pset_chunk(h_prop, rank, chunk_shape);
     if (h_err < 0)
       error("Error while setting chunk size (%llu, %llu) for field '%s'.",
-            (unsigned long long)chunk_shape[0],
-            (unsigned long long)chunk_shape[1], props.name);
+            chunk_shape[0], chunk_shape[1], props.name);
 
     /* Are we imposing some form of lossy compression filter? */
     if (lossy_compression != compression_write_lossless)
